refactor: Rename and restructure QueueCog for improved clarity and functionality
- Renamed `task_queue.py` to `queue.py` to better reflect its purpose and removed the old file.
- Updated `client.py` to load the renamed `QueueCog` correctly.
- Enhanced `QueueCog` with improved command documentation and pagination logic for displaying the download queue.
- Added detailed descriptions for command arguments and improved user feedback for queue management commands (show, clear, remove, pause, resume).
- Ensured all references to the old `task_queue` are updated to maintain functionality and consistency across the codebase.

These changes streamline the queue management functionality and improve the overall user experience with clearer command structures and responses.
