+++ src/boss_bot/bot/cogs/__init__.py
@@ -1,6 +1,6 @@
 """Discord cogs package for Boss-Bot."""
 
 from .downloads import DownloadCog
-from .task_queue import QueueCog
+from .queue import QueueCog
 
 __all__ = ["DownloadCog", "QueueCog"]
