+++ tests/test_bot/test_queue.py
@@ -18,7 +18,7 @@ from pytest_mock import MockerFixture
 from discord.ext import commands
 
 from boss_bot.bot.client import BossBot
-from boss_bot.bot.cogs.task_queue import QueueCog
+from boss_bot.bot.cogs.queue import QueueCog
 from boss_bot.core.core_queue import QueueItem, QueueStatus
 
 # Note: Using standardized fixtures from conftest.py:
